+++ Noodles/AggregationProperties/IAggregationProperty.cs
@@ -5,7 +5,7 @@ namespace Noodles.AggregationProperties
 {
     public interface IAggregationProperty
     {
-        bool Test(IEnumerable<double> values, IMeanCalculator calculator, AggregationProperty.AggregationTestTypes method = AggregationProperty.AggregationTestTypes.Default);
+        bool Test(IEnumerable<float> values, IMeanCalculator calculator, AggregationProperty.AggregationTestTypes method = AggregationProperty.AggregationTestTypes.Default);
         bool Test(IEnumerable<int> values, IMeanCalculator calculator, AggregationProperty.AggregationTestTypes method);
     }
 }
